Splitup public interface into multiple files
