+++ include/MiniLua/environment.hpp
@@ -1,5 +1,59 @@
-#ifndef ENVIRONMENT_H
-#define ENVIRONMENT_H
+#ifndef MINILUA_ENVIRONMENT_H
+#define MINILUA_ENVIRONMENT_H
+
+#include <iostream>
+#include <string>
+#include <unordered_map>
+
+#include "utils.hpp"
+
+namespace minilua {
+
+class Value;
+
+/**
+ * Represents the global environment/configuration for the 'Interpreter'.
+ *
+ * This contains things like global variables (including functions), etc.
+ */
+class Environment {
+    struct Impl;
+    owning_ptr<Impl> impl;
+
+public:
+    Environment();
+    Environment(const Environment&);
+    // can't use noexcept = default in older compilers (pre c++20 compilers)
+    // NOLINTNEXTLINE
+    Environment(Environment&&);
+    ~Environment();
+    auto operator=(const Environment&) -> Environment&;
+    // can't use noexcept = default in older compilers (pre c++20 compilers)
+    // NOLINTNEXTLINE
+    auto operator=(Environment &&) -> Environment&;
+    friend void swap(Environment&, Environment&);
+
+    // equivalent to the old env->populate_stdlib()
+    void add_default_stdlib();
+
+    void add(const std::string& name, Value value);
+    void add(std::string&& name, Value value);
+
+    void add_all(std::unordered_map<std::string, Value> values);
+    void add_all(std::initializer_list<std::pair<const std::string, Value>> values);
+
+    auto get(const std::string& name) -> Value&;
+
+    [[nodiscard]] auto size() const -> size_t;
+
+    friend auto operator==(const Environment&, const Environment&) noexcept -> bool;
+    friend auto operator!=(const Environment&, const Environment&) noexcept -> bool;
+    friend auto operator<<(std::ostream&, const Environment&) -> std::ostream&;
+};
+
+} // namespace minilua
+
+// TODO MOVE TO DIFFERENT FILE
 
 #include "val.hpp"
 
@@ -31,7 +85,7 @@ public:
     void populate_stdlib();
 };
 
-}
-}
+} // namespace rt
+} // namespace lua
 
 #endif
